# Speed-up profile management.