--- a/internal/proftest/proftest.go
+++ b/internal/proftest/proftest.go
@@ -18,13 +18,18 @@ package proftest
 
 import (
 	"encoding/json"
+	"flag"
 	"fmt"
 	"io"
 	"os"
 	"os/exec"
 	"regexp"
 	"testing"
+
+	_ "embed" // For embedding profiles needed by tests and benchmarks
 )
+
+var flagLargeProfile = flag.String("large_profile", "", "The name of a file that contains a profile to use in benchmarks. If empty, a profile of a synthetic program is used.")
 
 // Diff compares two byte arrays using the diff tool to highlight the
 // differences. It is meant for testing purposes to display the
@@ -76,7 +81,7 @@ func EncodeJSON(x interface{}) []byte {
 // Also tracks the number of times the error matches AllowRx in
 // NumAllowRxMatches.
 type TestUI struct {
-	T                 *testing.T
+	T                 testing.TB
 	Ignore            int
 	AllowRx           string
 	NumAllowRxMatches int
@@ -137,3 +142,25 @@ func (ui *TestUI) WantBrowser() bool {
 // SetAutoComplete is not supported by the test UI.
 func (ui *TestUI) SetAutoComplete(_ func(string) string) {
 }
+
+// LargeProfile returns a large profile that may be useful in benchmarks.
+//
+// If the flag --large_profile is set, the contents of the file
+// named by the flag are returned. Otherwise an embedded profile (~1.2MB)
+// for a synthetic program is returned.
+func LargeProfile(tb testing.TB) []byte {
+	tb.Helper()
+	if f := *flagLargeProfile; f != "" {
+		// Use custom profile.
+		data, err := os.ReadFile(f)
+		if err != nil {
+			tb.Fatalf("custom profile file: %v\n", err)
+		}
+		return data
+	}
+
+	return largeProfileData
+}
+
+//go:embed testdata/large.cpu
+var largeProfileData []byte