… internal/proftest/testdata/large.cpu